GUIScript: despite the moniker 'ID' and being 32 bits, controlIDs do *not* uniquely identify a control

+++ gemrb/core/ScriptEngine.h
@@ -30,7 +30,7 @@ namespace GemRB {
 
 class Point;
 
-typedef unsigned long ScriptingId;
+typedef unsigned int ScriptingId;
 typedef std::string ScriptingClassId;
 
 class ScriptingRefBase {
